--- src/mask/plugins/red-packet/types.ts
@@ -1,7 +1,0 @@
-export enum RequirementType {
-    Follow = 'Follow',
-    Like = 'Like',
-    Repost = 'Repost',
-    Comment = 'Comment',
-    NFTHolder = 'NFTHolder',
-}
