+++ src/modals/RedPacketModal/RedPacketContext.tsx
@@ -22,7 +22,7 @@ import { useProfileStoreAll } from '@/hooks/useProfileStore.js';
 import { useRedPacketThemes } from '@/hooks/useRedPacketThemes.js';
 import { EVMChainResolver } from '@/mask/bindings/index.js';
 import { RED_PACKET_DEFAULT_SHARES } from '@/mask/plugins/red-packet/constants.js';
-import { RequirementType } from '@/mask/plugins/red-packet/types.js';
+import { RequirementType } from '@/providers/red-packet/types.js';
 import type { FireflyRedPacketAPI } from '@/maskbook/packages/web3-providers/src/entry-types.js';
 
 export const redPacketRandomTabs = [
