+++ src/modals/RedPacketModal/RequirementsView.tsx
@@ -14,7 +14,7 @@ import { ActionButton } from '@/components/ActionButton.js';
 import { ClickableArea } from '@/components/ClickableArea.js';
 import { EMPTY_LIST } from '@/constants/index.js';
 import { Image } from '@/esm/Image.js';
-import { RequirementType } from '@/mask/plugins/red-packet/types.js';
+import { RequirementType } from '@/providers/red-packet/types.js';
 import { NonFungibleTokenCollectionSelectModalRef } from '@/modals/controls.js';
 import { RedPacketContext } from '@/modals/RedPacketModal/RedPacketContext.js';
